🦄 refactor(tests):  Credit - Auth abilities
Context: API is now protected by "abilities:api-access" middleware
- Using Sanctum::actingAs($user, ['api-access']); instead
$this->actingAs.
- Removed the use of $this->{function()} in benefit of
   native pest functions

Updates tests:
- Credit tests
