+++ .env.testing.example
@@ -75,3 +75,5 @@ WAREHOUSE_EMAIL_RECIPIENT=warehouse@example.com
 
 # Site logo URL used in transactional emails
 SITE_LOGO_URL=https://socomarca-frontend.vercel.app/assets/global/logo.png
+
+RANDOM_ERP_TOKEN=eyJhbGciOiJIUzI1NiIsInRfasdi9JkkVCJ9
