+++ .vscode/settings.json
@@ -38,18 +38,26 @@
   "cSpell.words": [
     "Assertable",
     "CRSD",
+    "CRSDCU",
+    "CRSDCV",
     "CRSDVU",
+    "CRSDVV",
     "Dispatchable",
     "idmaeedo",
+    "KOEN",
     "NVVBLV",
     "NVVFCV",
     "SOCOMARCA",
     "Spatie",
+    "SUEN",
     "superadmin",
     "supercategories",
     "supercategory",
     "tido",
     "transbank",
+    "uidxmaeedo",
+    "vabrdo",
+    "VALPO",
     "webpay",
     "Webpay"
   ]
